I thought about it last night, and the whole "arbitrary rules submodule was too complicated, so I deleted it and made the referee enforce the standard rules.
