--- rule_sets/__init__.py
@@ -1,1 +1,0 @@
-from rules import *
